Move text_encoders to encoders
