+++ tests/encoders/test_subword_tokenizer.py
@@ -6,13 +6,13 @@ import mock
 
 import six
 
-from torchnlp.text_encoders import subword_text_tokenizer
-from torchnlp.text_encoders.subword_text_tokenizer import encode
-from torchnlp.text_encoders.subword_text_tokenizer import decode
-from torchnlp.text_encoders.subword_text_tokenizer import _escape_token
-from torchnlp.text_encoders.subword_text_tokenizer import _unescape_token
-from torchnlp.text_encoders.subword_text_tokenizer import _ESCAPE_CHARS
-from torchnlp.text_encoders.subword_text_tokenizer import SubwordTextTokenizer
+from torchnlp.encoders import subword_text_tokenizer
+from torchnlp.encoders.subword_text_tokenizer import encode
+from torchnlp.encoders.subword_text_tokenizer import decode
+from torchnlp.encoders.subword_text_tokenizer import _escape_token
+from torchnlp.encoders.subword_text_tokenizer import _unescape_token
+from torchnlp.encoders.subword_text_tokenizer import _ESCAPE_CHARS
+from torchnlp.encoders.subword_text_tokenizer import SubwordTextTokenizer
 
 
 class TestTokenCounts(unittest.TestCase):
