+++ torchnlp/encoders/__init__.py
@@ -1,23 +1,23 @@
-from torchnlp.text_encoders.character_encoder import CharacterEncoder
-from torchnlp.text_encoders.delimiter_encoder import DelimiterEncoder
-from torchnlp.text_encoders.identity_encoder import IdentityEncoder
-from torchnlp.text_encoders.moses_encoder import MosesEncoder
-from torchnlp.text_encoders.reserved_tokens import COPY_INDEX
-from torchnlp.text_encoders.reserved_tokens import COPY_TOKEN
-from torchnlp.text_encoders.reserved_tokens import EOS_INDEX
-from torchnlp.text_encoders.reserved_tokens import EOS_TOKEN
-from torchnlp.text_encoders.reserved_tokens import PADDING_INDEX
-from torchnlp.text_encoders.reserved_tokens import PADDING_TOKEN
-from torchnlp.text_encoders.reserved_tokens import SOS_INDEX
-from torchnlp.text_encoders.reserved_tokens import SOS_TOKEN
-from torchnlp.text_encoders.reserved_tokens import UNKNOWN_INDEX
-from torchnlp.text_encoders.reserved_tokens import UNKNOWN_TOKEN
-from torchnlp.text_encoders.spacy_encoder import SpacyEncoder
-from torchnlp.text_encoders.static_tokenizer_encoder import StaticTokenizerEncoder
-from torchnlp.text_encoders.subword_encoder import SubwordEncoder
-from torchnlp.text_encoders.treebank_encoder import TreebankEncoder
-from torchnlp.text_encoders.whitespace_encoder import WhitespaceEncoder
-from torchnlp.text_encoders.text_encoder import TextEncoder
+from torchnlp.encoders.character_encoder import CharacterEncoder
+from torchnlp.encoders.delimiter_encoder import DelimiterEncoder
+from torchnlp.encoders.identity_encoder import IdentityEncoder
+from torchnlp.encoders.moses_encoder import MosesEncoder
+from torchnlp.encoders.reserved_tokens import COPY_INDEX
+from torchnlp.encoders.reserved_tokens import COPY_TOKEN
+from torchnlp.encoders.reserved_tokens import EOS_INDEX
+from torchnlp.encoders.reserved_tokens import EOS_TOKEN
+from torchnlp.encoders.reserved_tokens import PADDING_INDEX
+from torchnlp.encoders.reserved_tokens import PADDING_TOKEN
+from torchnlp.encoders.reserved_tokens import SOS_INDEX
+from torchnlp.encoders.reserved_tokens import SOS_TOKEN
+from torchnlp.encoders.reserved_tokens import UNKNOWN_INDEX
+from torchnlp.encoders.reserved_tokens import UNKNOWN_TOKEN
+from torchnlp.encoders.spacy_encoder import SpacyEncoder
+from torchnlp.encoders.static_tokenizer_encoder import StaticTokenizerEncoder
+from torchnlp.encoders.subword_encoder import SubwordEncoder
+from torchnlp.encoders.treebank_encoder import TreebankEncoder
+from torchnlp.encoders.whitespace_encoder import WhitespaceEncoder
+from torchnlp.encoders.text_encoder import TextEncoder
 
 __all__ = [
     'TextEncoder',
