+++ torchnlp/encoders/static_tokenizer_encoder.py
@@ -2,10 +2,10 @@ from collections import Counter
 
 import torch
 
-from torchnlp.text_encoders.reserved_tokens import EOS_INDEX
-from torchnlp.text_encoders.reserved_tokens import UNKNOWN_INDEX
-from torchnlp.text_encoders.reserved_tokens import RESERVED_ITOS
-from torchnlp.text_encoders.text_encoder import TextEncoder
+from torchnlp.encoders.reserved_tokens import EOS_INDEX
+from torchnlp.encoders.reserved_tokens import UNKNOWN_INDEX
+from torchnlp.encoders.reserved_tokens import RESERVED_ITOS
+from torchnlp.encoders.text_encoder import TextEncoder
 
 
 def _tokenize(s):
