+++ torchnlp/encoders/subword_encoder.py
@@ -1,10 +1,10 @@
 import torch
 
-from torchnlp.text_encoders.reserved_tokens import EOS_INDEX
-from torchnlp.text_encoders.reserved_tokens import UNKNOWN_INDEX
-from torchnlp.text_encoders.reserved_tokens import RESERVED_ITOS
-from torchnlp.text_encoders.subword_text_tokenizer import SubwordTextTokenizer
-from torchnlp.text_encoders.text_encoder import TextEncoder
+from torchnlp.encoders.reserved_tokens import EOS_INDEX
+from torchnlp.encoders.reserved_tokens import UNKNOWN_INDEX
+from torchnlp.encoders.reserved_tokens import RESERVED_ITOS
+from torchnlp.encoders.subword_text_tokenizer import SubwordTextTokenizer
+from torchnlp.encoders.text_encoder import TextEncoder
 
 
 class SubwordEncoder(TextEncoder):
